Add files via upload
These changes implement 2 major changes:

The first is the addition of Density: The way density works currently is that it is a hard coded array of integers (1-10) that multiplies the number of users to be added to each antenna. This is bad because it creates far too many users, and a better density function is needed, but for now this implementation demonstrates the concept of density.

 The second major change is the implementation of a 3rd "self-healing" algorithm. There was a need expressed in the last meeting for a version of the simulator that just ran, produced data, but did not self heal and this algorithm does that. This algorithm performs no self healing and instead just swaps users after they move to whatever base station that they moved to.

This new algorithm clashes with the density and attractiveness already implemented as one of the called functions is limited by a cap on users that I cannot find. So when the attractiveness or density of a base station is arbitrarily high this algorithm fails.